station-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
paho-mqtt==2.1.0
requests==2.32.0
pydantic==2.9.0
